requirements.txt
numpy
BioPython
scipy
